requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
httpx==0.28.1
pyyaml==6.0.3
jinja2==3.1.6
python-multipart==0.0.22
itsdangerous==2.2.0
sqlalchemy==2.0.48
aiosqlite==0.22.1
prometheus-client==0.24.1
apscheduler==3.10.4
# Optional async drivers (install the one matching your DATABASE_URL backend):
#   MySQL      : uv pip install aiomysql
#   PostgreSQL : uv pip install asyncpg
